feat: support TLS for all HTTP endpoints
Modify the IAM API server to enable TLS on its
gRPC, REST, and Metrics listeners.

This change introduces common command-line flags
(--tls-cert-file and --tls-key-file) for
specifying the TLS certificate and key. The server
will use these to serve all its endpoints over
HTTPS/GRPCS if the certificate and key are
provided.

The REST proxy and a new explicit Metrics server
(using promhttp) are now run in goroutines,
similar to the gRPC server, to allow concurrent
operation.