bugfix: use kubectl wrapper to interact with api
This commit resolves an issue with the `dev:setup` command that would
result in the setup process failing if the user's default kubeconfig
context was set to milo's kubeconfig. Any usage of kubectl in the
taskfile has been updated to use wrapper commands for kubectl that
automatically use the correct kubeconfig file.

A  new `kubectl` wrapper command has been added that can be used to
interact with the milo apiserver with the correct config file.

The test-infra taskfile was updated to use a stable version that
includes a `kubectl` wrapper command for using kubectl with the
test-infra cluster.