chore: add readiness check / improve validation
Creates a new HasSynced function that can be used to confirm whether the
validator's cache has been synced so this can be used by the apiserver
as a readiness check.

Also standardized on using the field error list to return errors back to
callers of the resource claims.

Adjusted the `claimingPermission` type to `claimingRules` to
disambiguate with IAM permissions.