feat: enable remote UID proxying feature gate
Our authorization system expects to use the UID to resolve the
authenticated user to their user profile in the system so we can confirm
they're a valid user. By default, the k8s apiserver does not include the
UID when proxying requests to aggregated apiservers.

From the docs on this feature flag:

> Enable the API server to accept UIDs (user IDs) via request header
> authentication. This will also make the kube-apiserver's API
> aggregator add UIDs via standard headers when forwarding requests to
> the servers serving the aggregated API.

See: https://kubernetes.io/docs/reference/command-line-tools-reference/feature-gates/